refactoring after moving validate_candidates
