+++ src/common_tasks/corpus_candidate_set.py
@@ -23,18 +23,6 @@ def validate_corpus_items(corpus_items: List[Dict]):
 
     return corpus_items
 
-@task()
-def validate_candidate_items(corpus_items: List[Dict]):
-    min_item_count = 1
-    expected_keys = ['ID', 'PUBLISHER']
-
-    assert len(corpus_items) >= min_item_count
-    assert all(set(expected_keys).difference(set(corpus_item.keys())) == set() for corpus_item in corpus_items)
-    assert all(isinstance(corpus_item['ID'], int) and corpus_item['ID'] != None for corpus_item in corpus_items)
-    assert all(isinstance(corpus_item['PUBLISHER'], str) and corpus_item['PUBLISHER'] != '' for corpus_item in corpus_items)
-
-    return corpus_items
-
 
 @task()
 def create_corpus_candidate_set_record(
